backend/requirements.txt
fastapi
uvicorn
motor
google-generativeai
python-dotenv
requests
google-api-python-client
langchain
langchain-google-genai
langchain-community
faiss-cpu
